% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2list.R
\name{array2list}
\alias{array2list}
\title{Convert an array(or matrix) to a list}
\usage{
array2list(X, MARGIN)
}
\arguments{
\item{X}{"array"}

\item{MARGIN}{"integer"}
}
\value{
"list"
}
\description{
Convert an array(or matrix) to a list
}
\details{
Also works with matrix.
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
