% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aT.x.DvecSigma4beta.inv.R
\name{aT.x.DvecSigma4beta.inv}
\alias{aT.x.DvecSigma4beta.inv}
\title{description}
\usage{
aT.x.DvecSigma4beta.inv(
  aT,
  Sigma.inv,
  diag.K.list,
  Xmats.delta.knots,
  P.type,
  p,
  q,
  q.i,
  n.par4knots
)
}
\value{

}
\description{
<description>
}
\details{
<details>
}
\note{
First version: Tue Feb 01 10:25:02 CET 2011;
      Current:       Tue Feb 01 10:25:08 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
