% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_forecast.R
\name{combination_forecast}
\alias{combination_forecast}
\title{Function to generate combination forecast}
\usage{
combination_forecast(
  predicted,
  ncomp = 2,
  accmat = NULL,
  real.error = NULL,
  tslist = TRUE,
  forecast_list = NULL,
  h = NULL,
  weights = TRUE,
  measure = NULL
)
}
\arguments{
\item{predicted}{prediction value matrix}

\item{ncomp}{number of algorithms use to combine}

\item{accmat}{function to compute forecast accuracy}

\item{real.error}{optional, matrix of MASE values for all algorithms for test data}

\item{tslist}{list of time series, as in the format of Mcomp object}

\item{forecast_list}{true forecast from different models}

\item{h}{length of forecast horizon}

\item{weights}{weighted by MASE}

\item{measure}{summary measure to be used mean or median}
}
\value{
list containing the forecasts and summaries
}
\description{
Select "n" number of algorithms based on minimum forecast error
}
