% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.K.R
\name{delta.K}
\alias{delta.K}
\title{Gradient for a diagonal matrix.}
\usage{
delta.K(diag.K, power)
}
\arguments{
\item{power}{"scalar".
The power for K to obtain the gradient for vec(K^(power)) w.r.t. diag(K)'.}

\item{K}{"matrix".
The diagonal matrix. p-by-p}
}
\value{
"matrix".
        The gradient,  pp-by-p
}
\description{
Give the gradient for a diagnoal matrix (K) w.r.t. its diagonal elements.
}
\details{
Details
}
\note{
First version: Sat Nov 06 16:31:01 CET 2010
      Current: 	 Mon Jan 10 18:02:02 CET 2011
}
\references{
The notes.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
