% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.sumlogdetP.R
\name{delta.sumlogdetP}
\alias{delta.sumlogdetP}
\title{<description>}
\usage{
delta.sumlogdetP(P.mats, q.i, Xmats.delta.knots.lst, priorArgs, n.par4knots)
}
\value{

}
\description{
<details>
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
