% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.sumqlogdetK .R
\name{delta.sumqlogdetK}
\alias{delta.sumqlogdetK}
\title{gradient for \sum q_i log|K_i|}
\usage{
delta.sumqlogdetK(q.i, diag.K)
}
\arguments{
\item{args}{"list" args$subset: the subsets of the diagonal K matrix.}
}
\value{
"matrix" 1-by-
}
\description{
gradient for \sum q_i log|K_i|
}
\note{
First version: Wed Jan 05 21:13:27 CET 2011;
      Current:       Wed Jan 05 21:13:37 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
