% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.vecPartiSigma4beta.inv.R
\name{delta.vecPartiSigma4beta.inv}
\alias{delta.vecPartiSigma4beta.inv}
\title{This is the gradient for a single partition of beta's variance.}
\usage{
delta.vecPartiSigma4beta.inv(diag.K_i, Sigma.inv, P_i)
}
\description{
This is the gradient for a single partition of beta's variance.
}
\note{
First version: Wed Jan 12 13:45:56 CET 2011; Current: Wed Jan 12 13:46:00 CET 2011.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
