% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensGradHess.R
\name{DensGradHess}
\alias{DensGradHess}
\title{A collection of gradient and Hessian for the log form of common densities.}
\usage{
DensGradHess(B, ..., type = "norm", grad = TRUE, Hess = TRUE)
}
\arguments{
\item{B}{"matrix".  The paramter that need to be added with a prior. The
gradient and hessian are calculated conditional on B. B should be always an
one-column matrix}

\item{...}{The arguments for the densities. The name should be matched
exactly.}

\item{type}{default norm}

\item{grad}{"logical" Should the gradient be computed?}

\item{Hess}{"logical" Should the Hessian be computed?}
}
\value{
"list". The gradient and Hessian (if required) matrices
}
\description{
The parameters after "..." should be matched exactly.
}
\note{
First version: Tue Mar 30 09:33:23 CEST 2010;
      Current:       Fri Mar 02 17:01:20 CET 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
