% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_prior.R
\name{density_prior}
\alias{density_prior}
\title{Densities for priors}
\usage{
density_prior(vecB, prior_type, log = TRUE, ...)
}
\arguments{
\item{vecB}{"Matrix".
q-by-1, the prameters to be added with a prior.}

\item{prior_type}{"Charactor".
Types of prior}

\item{log}{"Logical".
IF TRUE(default), the logarithm density is returned.}

\item{...}{Addtional arguments w.r.t. different priors.}
}
\value{
"Matrix" p-by-1.
}
\description{
Densities for priors.
}
\details{
A collection of density for common priors.
}
\note{
First version: Wed Mar 31 10:50:41 CEST 2010;
      Current:       Wed Sep 15 10:36:55 CEST 2010.

DEPENDS: mvtnorm
TODO: This function is used for the conditional model,  not the conjugate case.
}
\references{
g-Prior: Zellner(1986)
}
\author{
Feng Li, Dept. of Statistics, Stockholm University, Sweden.
}
