% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.hwang.R
\name{DGP}
\alias{DGP}
\alias{DGP.hwang}
\title{DATA GENERATING PROCESSES}
\usage{
DGP.hwang(
  n,
  q,
  Sigma,
  model,
  otherArgs = list(seed = NA, nTesting = NA),
  PlotData
)
}
\arguments{
\item{n}{"numeric".
No. of obs. to be generated}

\item{Sigma}{"matrix".
The variance covariance matrix.}

\item{model}{"character".
The model name to be use.
When the model is "spline",  you also need to supply paramters,  see otherArgs.}

\item{otherArgs}{"list".
        Other arguemnts need to pass to the function.
        otherArgs$seed: Set the seed. If this vaule is unset, use a random
seed.
        otherArgs$xi:
        otherArgs$withOnt:
        otherArgs$spline.type:}

\item{PlotData}{"logical"
If need to plot the data out.}
}
\value{
"list"
}
\description{
A collection of DGPs for different models from Hwang's example.
}
\details{
The model can be extended.
}
\note{
First version: Tue Mar 30 20:11:06 CEST 2010;
      Current:       Tue Jul 31 13:05:02 CEST 2012.
DEPENDS: mvtnorm
}
\references{
Hwang 2000
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
