% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.surface.R
\name{DGP.surface}
\alias{DGP.surface}
\title{A DGP process where the true model nests in the fitted model}
\usage{
DGP.surface(n, p, q.o, q.s, Sigma, splineArgs, splineArgs.crl, PlotDGP = FALSE)
}
\value{

}
\description{
DGP Surface nested
}
\details{
A DGP process where the true model nests in the fitted model
}
\note{
First version: Mon Apr 11 09:18:38 CEST 2011;Current:       Tue Jul 31 08:54:18 CEST 2012.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
