% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{diwishart}
\alias{diwishart}
\title{##' Density for the inverse wishart distribution}
\usage{
diwishart(X, df, V, log = TRUE)
}
\arguments{
\item{V}{Location matrix}
}
\value{

}
\description{
Details.
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
