% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixmorm.R
\name{dmixnorm.ts}
\alias{dmixnorm.ts}
\title{The conditional density for AR-type mixture of normals}
\usage{
dmixnorm.ts(y, means.ar.par.list, sigmas.list, weights, log = FALSE)
}
\arguments{
\item{y}{An `n`-length vector. The time series.}

\item{means.ar.par.list}{An `k`-length list. Each element in the list consists the
coefficients in the AR process.}

\item{weights}{A `k`-length vector. The weight in each component.}

\item{log}{Logical; If TRUE, the log density is returned.}
}
\description{
The conditional density for AR-type mixture of normals
}
