% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features.df.m1}
\alias{features.df.m1}
\title{features.df.m1}
\format{
dataframe features calculated on M1 competition data
}
\usage{
features.df.m1
}
\description{
features calculated on M1 competition data
}
\keyword{datasets}
