% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features.df}
\alias{features.df}
\title{features.df}
\format{
A data frame with 645 rows and 25 variables
}
\usage{
features.df
}
\description{
features calculated on M3 competition data
}
\examples{
data(features.df)
head(features.df)
}
\keyword{datasets}
