% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featuresM1Y}
\alias{featuresM1Y}
\title{featuresM1Y}
\format{A data frame with 181 rows and 8 variables
\describe{
\item{trend}{strength of trend}
\item{ur_pp}{test statistic of Phillip-Perron test }
\item{spikiness}{Spikiness}
\item{beta}{parameter estimate of beta in ETS(A,A,N)}
\item{alpha}{parameter estimate of alpha in ETS(A,A,N)}
\item{diff1y_acf1}{first ACF value of the differenced series}
\item{linearity}{strength of linearity in a time series}
\item{curvature}{strength of curvature}
\item{N}{length of the series}
}}
\usage{
featuresM1Y
}
\description{
Six time series features calculated on M1 competition data
}
\examples{
data(featuresM1Y)
head(featuresM1Y)
}
\keyword{datasets}
