% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fformpp.model}
\alias{fformpp.model}
\title{fformpp.model}
\format{list containing fitted model and spline arguments}
\usage{
fformpp.model
}
\description{
fitted model from fit_fformpp
}
\keyword{datasets}
