% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fformpp.R
\name{fit_surfreg}
\alias{fit_surfreg}
\title{Function to fit Efficient Bayesian Surface Regression Models.
This function is written based on the example codes in movingknots
package by Dr. Feng Li. Please see https://github.com/feng-li/movingknots for more details.}
\usage{
fit_surfreg(feamat, accmat, sknots = 2, aknots = 2, fix.s = 0,
  fix.a = 0, fix.shrinkage, fix.covariance = 0, fix.coefficients = 0,
  n.iter = 100, knot.moving.algorithm = "Random-Walk",
  ptype = c("identity", "identity", "identity"), prior.knots)
}
\arguments{
\item{feamat}{matrix of features, rows corresponds to each time series,
columns coresponds to features}

\item{accmat}{matrix of forecast errors from each method,
rows represent time series, columns represent forecast algorithms}

\item{sknots}{the dimension of knots for surface, default 2}

\item{aknots}{no. of knots used in each covariates for the additive part, default 2}

\item{fix.s}{number of knots to be fixed in the surface components, 0 means all are
updated}

\item{fix.a}{number of knots to be fixed in the additive components, default is 0 which means all are updated}

\item{fix.shrinkage, }{number of shrinkage covariates not to be updated, defalut is, 1:p,}

\item{fix.covariance, }{number of knots to be fixed in the covariance, default is 0, all are updated}

\item{fix.coefficients, }{number of knots to be fixed in the coefficients, default is 0, all are updated}

\item{n.iter}{number of ierations}

\item{knot.moving.algorithm, }{select either "KStepNewton" or "Random-Walk", to fasten the
code use Random-Walk}

\item{ptype}{For fixing gprior, This could be c("X'X", "identity", "identity") or c("identity", "identity", "identity")}

\item{prior.knots}{this could be n, log(n) or 1 # to set priors for knots}
}
\value{
returns a list contanining the fitted model and arguments for splines
}
\description{
Function to fit Efficient Bayesian Surface Regression Models.
This function is written based on the example codes in movingknots
package by Dr. Feng Li. Please see https://github.com/feng-li/movingknots for more details.
}
