% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDiagnosis.hwang.R
\name{FitDiagnosis.hwang}
\alias{FitDiagnosis.hwang}
\title{Diagnosis if the spline model}
\usage{
FitDiagnosis.hwang(
  x.lst,
  Y,
  OUT.Params,
  Data.gen,
  logpost.fun.name,
  splineArgs,
  Params_Transform,
  crossvalid.struc,
  burn.in,
  criterion = c("LOSS", "KL", "L2"),
  hwang.model
)
}
\description{
Diagnosis if the spline model
}
\note{
First version: Tue Aug 07 16:55:03 CEST 2012;
      Current: Tue Aug 07 16:54:52 CEST 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
