% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDiagnosis.R
\name{FitDiagnosis}
\alias{FitDiagnosis}
\title{Diagnosis if the spline model}
\usage{
FitDiagnosis(
  x,
  Y,
  OUT.Params,
  Data.gen,
  logpost.fun.name,
  splineArgs,
  splineArgs.gen,
  Params_Transform,
  crossvalid.struc,
  burn.in,
  criterion = c("LOSS", "KL", "L2")
)
}
\description{
This function is only applied to the moving knots model.
}
