% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldMat.R
\name{foldMat}
\alias{foldMat}
\title{Fold a matrix}
\usage{
foldMat(Mat, nfolds, byrow = TRUE)
}
\value{

}
\description{
<details>
}
\note{
First version: Thu Jan 27 20:09:56 CET 2011;
      Current:       Thu Jan 27 20:09:59 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
