% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast.error.m1}
\alias{forecast.error.m1}
\title{forecast.error.m1}
\format{
dataframe features calculated on M1 competition data
}
\usage{
forecast.error.m1
}
\description{
forecast error calculated on M1 data
}
\keyword{datasets}
