% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast.error}
\alias{forecast.error}
\title{forecast.error}
\format{A data frame with 181 rows with 7 variables
\describe{
\item{ets}{ets algorithm in the forecast package}
\item{arima}{auto.arima algorithm in the forecast package}
\item{rw}{random walk}
\item{rwd}{random walk with drift}
\item{wn}{white noise process}
\item{theta}{theta approach}
\item{nn}{neural network approach}
}}
\usage{
forecast.error
}
\description{
forecast.error
}
\keyword{datasets}
