% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{transform.features}
\alias{transform.features}
\title{Function to transform features}
\usage{
\method{transform}{features}(feature, transformation)
}
\arguments{
\item{feature}{feature variable you need to transform}

\item{transformation}{the transformation, "logit"}
}
\value{
vector of transformed values of the feature
}
\description{
Function to transform features
}
