% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_logpost.R
\name{glm_logpost}
\alias{glm_logpost}
\title{The conditional and joint log posterior function}
\usage{
glm_logpost(Y, x, Params, callParam, splineArgs, priorArgs, Params_Transform)
}
\arguments{
\item{Params}{Params$xi:
Parmas$K:}

\item{callParam}{callParam$id: It should be able to obtain conditianl posterior or joint
posterior or just likelihood.
callParam$subset: If the paramater set is larg, we may only update a subset of
them.}

\item{splineArgs}{"list".}

\item{priorArgs}{priorArgs$prior_type:
priorArgs$n0:
priorArgs$S0:
priorArgs$mu:
priorArgs$M:
priorArgs$mu0:
priorArgs$Sigma0:
priorArgs$ka0:
priorArgs$ka.mu0:
priorArgs$ka.Sigma0:
priorArgs$Sigma.mu0:
priorArgs$Sigma.Sigma0:}

\item{ParamsTransArgs}{"list"}
}
\value{
"scalar".
}
\description{
Details.
}
\note{
First version: Sat Nov 06 22:23:57 CET 2010;
      Current:       Wed Dec 29 15:55:51 CET 2010.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
