% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{glogit}
\alias{glogit}
\title{logit transformation}
\usage{
glogit(x, a, b)
}
\arguments{
\item{x}{variable name}

\item{a}{lower boundary for transformation}

\item{b}{upper boundary for transformation}
}
\value{
a vector of logit transformed feature values
}
\description{
logit transformation
}
