% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.x.deriv_link.R
\name{grad.x.deriv_link}
\alias{grad.x.deriv_link}
\title{D[vecf(g(x)]/D[vecg(x)'] %*% D[vec g(x)]/D[vec x']}
\usage{
grad.x.deriv_link(grad, Param, link)
}
\arguments{
\item{grad}{"matrix" gradient}

\item{par}{"matrix" should always be a matrix}
}
\value{

}
\description{
Just the chain rule
}
\note{
First version: Sun Feb 20 15:03:34 CET 2011;
      Current:       Sun Feb 20 15:03:43 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
