% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagAlg.R
\name{\%d*\%}
\alias{\%d*\%}
\title{Fast matrix algebra with diagonal matrices.}
\usage{
d \%d*\% M
}
\arguments{
\item{d}{"vector" Diagonal elements form a diagonal matrix.}

\item{M}{"matrix" Dense matrix that is conformable to matrix diag(d).}
}
\description{
Fast matrix algebra with diagonal matrices.
}
\note{
First version: Tue Nov 09 11:12:05 CET 2010;
      Current:       Sat Oct 04 17:38:57 CST 2014.
Thanks for Bertil's naming
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
