% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.b2beta.R
\name{idx.b2beta}
\alias{idx.b2beta}
\title{Make indices from b to beta.}
\usage{
idx.b2beta(p, q, q.i)
}
\value{

}
\description{
<details>
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
