% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.beta2b.R
\name{idx.beta2b}
\alias{idx.beta2b}
\title{Convert indices from beta to b}
\usage{
idx.beta2b(p, q, q.i)
}
\value{

}
\description{
<details>
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
