% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_forecast.R
\name{individual_forecast}
\alias{individual_forecast}
\title{Function to generate forecasts from minimum error}
\usage{
individual_forecast(predicted, accmat = NULL, real.error = NULL,
  tslist = TRUE, forecast_list = NULL, h = NULL)
}
\arguments{
\item{predicted}{predicted forecast errors from predict_fformpp}

\item{accmat}{function to compute forecast accuracy}

\item{real.error}{optional, matrix of MASE values for all algorithms for test data}

\item{tslist}{list of time series, as in the format of Mcomp object}

\item{forecast_list}{true forecast from different models}

\item{h}{length of forecast horizon}
}
\description{
Function to generate forecasts from minimum error
}
