% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_check_boundary.R
\name{knots_check_boundary}
\alias{knots_check_boundary}
\title{Description}
\usage{
knots_check_boundary(P4X, method = "singular")
}
\arguments{
\item{method}{"character"}

\item{...}{Input with respect to different methods.}
}
\value{
"character" "ok": the knots are at good position. "bad": bad knots locations.
}
\description{
Details.
}
\note{
First version: Tue Oct 12 22:15:18 CEST 2010;
      Current:       Tue Oct 12 22:15:35 CEST 2010.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
