% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_list2mat.R
\name{knots_list2mat}
\alias{knots_list2mat}
\title{Convert list of knots into a n-by-1 matrix.}
\usage{
knots_list2mat(knots.list, byrow = TRUE)
}
\arguments{
\item{knots.list}{List}

\item{byrow}{"logical"}
}
\value{
Matrix
}
\description{
Converted byrow by default.
}
\note{
First version: Wed Feb 02 13:01:00 CET 2011;
      Current:       Wed Feb 02 13:01:07 CET 2011.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
