% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_mat2list.R
\name{knots_mat2list}
\alias{knots_mat2list}
\title{convert the knots matrix into the knots list}
\usage{
knots_mat2list(knots.mat, splineArgs)
}
\value{

}
\description{
<details>
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
