% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_subsetsIdx.R
\name{knots_subsetsIdx}
\alias{knots_subsetsIdx}
\title{This function is to find the locations of subsets in the knots matrix}
\usage{
knots_subsetsIdx(knots.subsets, k, m)
}
\arguments{
\item{knots.subsets}{"list"}

\item{k}{"integer"
No. of knots in the knots matrix.}

\item{m}{"integer"
Dimension of knots in the knots matrix.}
}
\value{
"list"
}
\description{
This function is to find the locations of subsets in the knots matrix
}
\note{
First version: Sat Oct 02 15:51:01 CEST 2010;
      Current:       Sat Oct 02 15:51:10 CEST 2010.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
