% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStepNewtonMove.R
\name{KStepNewtonMove}
\alias{KStepNewtonMove}
\title{This is the simple Newton move for spline models without dimession changes.}
\usage{
KStepNewtonMove(
  param.cur,
  gradhess.fun.name,
  KStep,
  Params,
  hessMethod,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\description{
This is the simple Newton move for spline models without dimession changes.
}
