% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_IWishart.R
\name{linear_IWishart}
\alias{linear_IWishart}
\title{Calculate the posterior df and location matrix V from the inverse Wishart distribution.}
\usage{
linear_IWishart(Params, Y, x0, splineArgs, priorArgs, Params_Transform)
}
\description{
Calculate the posterior df and location matrix V from the inverse Wishart distribution.
}
