% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_post4coef.R
\name{linear_post4coef}
\alias{linear_post4coef}
\title{Direct sample the coefficients from normal distribution.}
\usage{
linear_post4coef(
  Y,
  x0,
  OUT.Params,
  crossvalid.struc,
  nCross,
  nIter,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{Y}{NA}

\item{x0}{NA}

\item{OUT.Params}{NA}

\item{crossvalid.struc}{NA}

\item{nCross}{NA}

\item{nIter}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transform}{NA}
}
\value{
NA
}
\description{
Details are available from the paper.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
