% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior.R
\name{log_prior}
\alias{log_prior}
\title{logarithm density for priors}
\usage{
log_prior(B, priorArgs)
}
\arguments{
\item{B}{"matrix".
The paramter that need to be added with a prior. The prior density is
calculated conditional on B. B should be always an one-column matrix,}

\item{priorArgs}{"list".
when
priorArgs$prior_type: is set to "mvnorm", you have to provide
priorArgs$mean: "matrix", the mean of parameter, mu0 should be always an
one-column matrix;
priorArgs$covariance: "matrix", the covariance matrix. A g-prior can be
constructed by setting it to X'X, where X is the covariates matrix.;
priorArgs$shrinkage: "numeric", the shrinkage for the covariance.}
}
\value{
"scalor". The log density of priors.
}
\description{
A collection of log densities for common priors.
}
\details{
The parameters after "..." should be matched exactly.
}
\note{
First version: Tue Mar 30 09:33:23 CEST 2010;
      Current:       Wed Sep 15 14:39:01 CEST 2010.

DEPENDS: mvtnorm::dmvnorm
TODO:
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
