% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogPredScore.R
\name{LogPredScore}
\alias{LogPredScore}
\title{This is the LPDS}
\usage{
LogPredScore(
  Y,
  x,
  logpost.fun.name,
  crossvaid.struc,
  splineArgs,
  priorArgs,
  OUT.Params,
  Params_Transform,
  burn.in,
  LPDS.sampleProp
)
}
\value{

}
\description{
Details.
}
\note{
First version: Tue Nov 30 23:18:18 CET 2010;
      Current:       Tue Nov 30 23:18:25 CET 2010.
TODO: Don't use the full draws when nIter is large.
      Still doubt the nseLPDS
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
