% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.knotsPriVar.R
\name{make.knotsPriVar}
\alias{make.knotsPriVar}
\title{set the prior variance of the knots.}
\usage{
make.knotsPriVar(x, splineArgs)
}
\value{

}
\description{
<details>
}
\note{
First version: Thu Feb 17 10:22:12 CET 2011;
      Current:       Thu Feb 17 10:22:19 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
