% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stepsize.R
\name{make_stepsize}
\alias{make_stepsize}
\title{Make step size for SGLD}
\usage{
make_stepsize(steprange, n, args)
}
\value{
NA
}
\description{
Step size is also known as the learning rate in stochastic gradient descent
algorithms.
}
\references{
NA
}
\author{
Feng Li, Central University of Finance and Economics.
}
