% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.delta.xi.R
\name{Mat.delta.xi}
\alias{Mat.delta.xi}
\title{<description>}
\usage{
Mat.delta.xi(Mat, delta.knots, n, p, q.i)
}
\value{

}
\description{
<details>
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
