% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.AT.k.I.x.K.x.delta.knots.R
\name{Mat.x.AT.k.I.x.K.x.delta.knots}
\alias{Mat.x.AT.k.I.x.K.x.delta.knots}
\title{<description>}
\usage{
Mat.x.AT.k.I.x.K.x.delta.knots(Mat, A, delta.knots, p, q, q.i)
}
\arguments{
\item{Mat}{"matrix" Sigma4beta.tilde}
}
\value{

}
\description{
<details>
}
\note{
First version: Thu Feb 10 10:19:02 CET 2011;
      Current:       Thu Feb 10 10:19:08 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
