% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.delta.vecSigma4beta.inv.R
\name{Mat.x.delta.vecSigma4beta.inv}
\alias{Mat.x.delta.vecSigma4beta.inv}
\title{Compute a matrix to pre-multiply the gradient for vec beta's inverse variance}
\usage{
Mat.x.delta.vecSigma4beta.inv(Mat, delta4K.list, p, q, q.i)
}
\arguments{
\item{args}{"list" args$subset: subset of the shrinkage}
}
\value{

}
\description{
<details>
}
\note{
First version: Wed Jan 19 16:37:25 CET 2011;
      Current:       Wed Jan 19 16:37:34 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
