% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.DvecA.k.P_stp1.R
\name{Mat.x.DvecA.k.P_stp1}
\alias{Mat.x.DvecA.k.P_stp1}
\title{Perform a dense matrix multiplying by Dev[vec[A %*% P_i]]/Dev[vecX_i'] %*% (I_q_i^2 +
K_qi,qi)}
\usage{
Mat.x.DvecA.k.P_stp1(Mat, A, p, q, q_i)
}
\arguments{
\item{Mat}{"Matrix"}

\item{A}{"Matrix" Spares}
}
\value{
"Matrix"
}
\description{
Perform a dense matrix multiplying by Dev[vec[A %*% P_i]]/Dev[vecX_i'] %*% (I_q_i^2 +
K_qi,qi)
}
\note{
First version: Thu Jan 27 17:41:38 CET 2011;
      Current:       Thu Jan 27 17:41:45 CET 2011.
}
\references{
Notes
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
