% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.DvecA.k.P_stp2.R
\name{Mat.x.DvecA.k.P_stp2}
\alias{Mat.x.DvecA.k.P_stp2}
\title{<description>}
\usage{
Mat.x.DvecA.k.P_stp2(Mat, Xmats.delta.knots.i, p, q, q_i)
}
\arguments{
\item{Mat}{"matrix"}

\item{X_i}{"matrix" the matrix used for obtaining P_i matrix}

\item{delta_i}{"matrix" the matrix for the gradient}
}
\value{
"matrix"
}
\description{
<details>
}
\note{
First version: Fri Jan 28 11:30:57 CET 2011;
      Current:       Fri Jan 28 11:31:06 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
