% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.DvecSigma.inv.k.XTX.R
\name{Mat.x.DvecSigma.inv.k.XTX}
\alias{Mat.x.DvecSigma.inv.k.XTX}
\title{<description>}
\usage{
Mat.x.DvecSigma.inv.k.XTX(Mat, Sigma.inv, X.delta.knots.lst, p, q, q.i)
}
\value{

}
\description{
<details>
}
\note{
First version: Tue Feb 01 10:25:02 CET 2011;
      Current:       Tue Feb 01 10:25:08 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
