% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.trajectory.R
\name{MCMC.trajectory}
\alias{MCMC.trajectory}
\title{Trajectory MCMC.}
\usage{
MCMC.trajectory(iIter, nIter, iCross, OUT.accept.probs, interval = 0.1)
}
\arguments{
\item{interval}{e.g. 10%}
}
\value{

}
\description{
Details.
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
