% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.grid.R
\name{mesh.grid}
\alias{mesh.grid}
\title{Same as expand.grid function but return a matrix.}
\usage{
mesh.grid(x1, x2 = x1)
}
\value{

}
\description{
Details.
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
