% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPropWithIWishart.R
\name{MHPropWithIWishart}
\alias{MHPropWithIWishart}
\title{Random walk Metropolis–Hastings algorithm for Sigma}
\usage{
MHPropWithIWishart(
  param.cur,
  logpost.fun.name,
  Params,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\value{

}
\description{
Details.
}
\note{
First version: Mon Nov 15 18:23:37 CET 2010;
      Current:       Wed Feb 23 09:26:07 CET 2011.
DEPENDS:
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
