% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPropWithKStepNewton.R
\name{MHPropWithKStepNewton}
\alias{MHPropWithKStepNewton}
\title{Metropolis–Hastings algorithm with K-step Newton method for the spline model.}
\usage{
MHPropWithKStepNewton(
  param.cur,
  gradhess.fun.name,
  logpost.fun.name,
  nNewtonStep,
  Params,
  hessMethod,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  prop.df,
  Params_Transform
)
}
\value{

}
\description{
Details.
}
\note{
First version: Thu Feb 17 14:03:14 CET 2011;
      Current:       Wed Feb 23 14:43:35 CET 2011.
DEPENDS: mvtnorm
TODO: replace the old multivariate t functions to mvtnorm functions
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
