% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.matrix.R
\name{P.matrix}
\alias{P.matrix}
\title{Generate the P matrices for the moving knots model}
\usage{
P.matrix(X, q.i, priorArgs)
}
\arguments{
\item{X}{"matrix" the design matrix generated by d.matrix() function.}

\item{q}{"numeric" the dimensions for the P matrices}

\item{args}{"list"
args$P.type: "character" the type for the P matrices}
}
\value{
"list"
        Containing the P matrices.
}
\description{
Generate the P matrices for the moving knots model
}
\note{
First version: Sun Jan 02 15:50:01 CET 2011;
      Current:       Sun Jan 02 15:50:09 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
