% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.transform2.R
\name{par.transform2}
\alias{par.transform2}
\title{Transform the parameters from the original scale to the new scale (the final scale) wrt the linkages}
\usage{
par.transform2(par, method)
}
\value{

}
\description{
Reverse step of par.transfrom()
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
