% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parLinkFun.R
\name{parLinkFun}
\alias{parLinkFun}
\title{<short tile>}
\usage{
parLinkFun(mu, linkArgs)
}
\value{
"one-col-matrix" of the same dimension as the linear predictor
}
\description{
Transform the mean function in the GLM framework.
}
\details{
<details>
}
\note{
Created: Thu May 10 13:09:08 CEST 2012;
      Current: Thu May 10 13:09:14 CEST 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
