% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parMeanFun.R
\name{parMeanFun}
\alias{parMeanFun}
\title{GLM mean function}
\usage{
parMeanFun(X, beta, linkArgs)
}
\arguments{
\item{X}{"matrix"}

\item{beta}{"one-col-matrix"}

\item{link}{"character" Type of link function}
}
\value{
"one-col-matrix" of the same dimension as the linear predictor
}
\description{
Transform the mean function in the GLM framework.
}
\note{
Created: Sun Mar 04 14:45:22 CET 2012; Current: Wed Sep 30 12:34:04 CST 2015.
}
\references{
NA
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
