% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkMetropolis.R
\name{RandomWalkMetropolis}
\alias{RandomWalkMetropolis}
\title{Random Walk Metropolis}
\usage{
RandomWalkMetropolis(
  param.cur,
  logpost.fun.name,
  Params,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\value{

}
\description{
Random walk Metropolis algorithm.
}
\details{
<details>
}
\note{
Created: Fri Aug 31 13:44:15 CEST 2012;
      Current: Fri Aug 31 13:44:21 CEST 2012.
}
\references{
Li Villani 2012
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
