% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{riwishart}
\alias{riwishart}
\title{Description}
\usage{
riwishart(df, V)
}
\value{

}
\description{
Details.
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
