% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixmorm.R
\name{rmixnorm}
\alias{rmixnorm}
\title{Random variables from normal mixture}
\usage{
rmixnorm(n, means, sigmas, weights)
}
\arguments{
\item{n}{An positive integer. Numbers of samples to be generated.}

\item{means}{A `q-by-k` matrix. Mean value vector within each component in total k
components.}

\item{sigmas}{A `q-by-q-by-k` array.  Variance covariance matrix with in each
component.}

\item{weights}{A `k`-length vector.  The weight in each component.}
}
\value{
An `n-by-q` matrix
}
\description{
Generate random variables from univariate and multivariate mixture normal distribution
}
\details{
Random variables from mixture of normals.
}
\references{
Villani et al 2009
}
\author{
Feng Li, Central University of Finance and Economics.
}
