% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixmorm.R
\name{rmixnorm.ts}
\alias{rmixnorm.ts}
\title{AR-type random variables from mixture of normals}
\usage{
rmixnorm.ts(n, means.ar.par.list, sigmas.list, weights, yinit = 0)
}
\arguments{
\item{n}{a positive integer.  Number of samples.}

\item{means.ar.par.list}{parameters in AR(p) within each mixing component}

\item{sigmas.list}{variance list}

\item{weights}{weight in each list}

\item{yinit}{initial values}
}
\value{
vector of n follows a mixture distribution
}
\description{
Simulating AR type random variables from mixture of normals
}
\details{
This function simulates random samples from a finite mixture of Gaussian distribution
    where the mean from each components are AR(p) process.
}
\references{
Li 2010 JSPI
}
\author{
Feng Li, Central University of Finance and Economics.
}
