% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.all.R
\name{save.all}
\alias{save.all}
\title{Save all the objects to a folder.}
\usage{
save.all(save.output, ModelDescription)
}
\description{
Save all the objects to a folder.
}
\note{
First version: Sun Sep 02 16:10:36 CEST 2012;
      Current: Sun Sep 02 16:10:42 CEST 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
