% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.crossvalid.R, R/set_crossvalid.R
\name{set.crossvalid}
\alias{set.crossvalid}
\title{Setup cross-validation}
\usage{
set.crossvalid(nObs, crossValidArgs)

set.crossvalid(nObs, crossValidArgs)
}
\arguments{
\item{n.obs}{number of observations}

\item{crossvalidArgs}{arguments for cross validation}
}
\value{
list

list
}
\description{
Setup cross-validation

Setup cross-validation
}
\note{
First: Thu May 03 14:36:26 CEST 2012; Current: Thu Aug 27 19:19:43 CST 2015

First: Thu May 03 14:36:26 CEST 2012; Current: Thu Aug 27 19:19:43 CST 2015
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.

Feng Li, Department of Statistics, Stockholm University, Sweden.
}
