% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGLD.R
\name{SGLD}
\alias{SGLD}
\title{Stochastic MCMC using Stochastic gradient Langevin dynamics}
\usage{
SGLD(param.cur, logpost.fun.name, Params, Y, x0, callParam, splineArgs,
  priorArgs, algArgs, Params_Transform)
}
\arguments{
\item{param.cur}{NA}

\item{logpost.fun.name}{MA}

\item{Params}{NA}

\item{Y}{NA}

\item{x0}{NA}

\item{callParam}{NA}

\item{splineArgs}{NA}

\item{priorArgs}{NA}

\item{Params_Transfor}{NA}
}
\value{
NA
}
\description{
<description>
}
\references{
Ma, Chen, & Fox 2015 A Complete Recipe for Stochastic Gradient MCMC.
}
\author{
Feng Li, Central University of Finance and Economics.
}
