% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma4betaFun.R
\name{Sigma4betaFun}
\alias{Sigma4betaFun}
\title{Calculate the variance for the prior of coefficients (beta) in the moving knots model}
\usage{
Sigma4betaFun(diag.K, Sigma, P.mats, inverse = FALSE)
}
\description{
Calculate the variance for the prior of coefficients (beta) in the moving knots model
}
