% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Print object.size}
\usage{
size(x, unit = "auto")
}
\value{

}
\description{
<details>
}
\note{
First version: Mon Jan 17 20:05:44 CET 2011;
      Current:       Mon Jan 17 20:05:51 CET 2011.
TODO: allow multiple inputs
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
