% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vech.R
\name{vech}
\alias{vech}
\title{calculating a vector}
\usage{
vech(X, diag = TRUE)
}
\arguments{
\item{X}{data matrix}

\item{diag}{diagonal values}
}
\value{
out a matrix
}
\description{
Vech function
}
\details{
Details.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
