% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.x.delta.xi.R
\name{X.x.delta.xi}
\alias{X.x.delta.xi}
\title{Preform X multiply Dev xi}
\usage{
X.x.delta.xi(X, delta.knots, q.i)
}
\description{
<details>
}
\note{
First version: Thu Feb 10 11:09:46 CET 2011;
      Current:       Thu Feb 10 11:09:53 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
