% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmats.x.delta.xi.R
\name{Xmats.x.delta.xi}
\alias{Xmats.x.delta.xi}
\title{Preform X_i multiply Dev xi}
\usage{
Xmats.x.delta.xi(X.mats, delta.knots, q.i, P.type)
}
\description{
<details>
}
\note{
First version: Thu Feb 10 11:09:46 CET 2011;
      Current:       Thu Feb 10 11:09:53 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
