% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aT.x.DvecSigma4beta.inv.R
\name{aT.x.DvecSigma4beta.inv}
\alias{aT.x.DvecSigma4beta.inv}
\title{<description>}
\usage{
aT.x.DvecSigma4beta.inv(
  aT,
  Sigma.inv,
  diag.K.list,
  Xmats.delta.knots,
  P.type,
  p,
  q,
  q.i,
  n.par4knots
)
}
\description{
<description>
}
