% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.densParams.R
\name{convert.densParams}
\alias{convert.densParams}
\title{Convert the parameters corresponding to different transformations}
\usage{
convert.densParams(mean, var, linkage)
}
\description{
Convert the parameters corresponding to different transformations
}
