% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDiagnosis.R
\name{FitDiagnosis}
\alias{FitDiagnosis}
\title{Diagnosis if the spline model}
\usage{
FitDiagnosis(
  x,
  Y,
  OUT.Params,
  Data.gen,
  logpost.fun.name,
  splineArgs,
  splineArgs.gen,
  Params_Transform,
  crossvalid.struc,
  burn.in,
  criterion = c("LOSS", "KL", "L2")
)
}
\value{

}
\description{
Diagnosis if the spline model
}
\details{
This function is only applied to the moving knots model.
}
\note{
First version: Tue Aug 07 16:55:03 CEST 2012;
      Current: Tue Aug 07 16:54:52 CEST 2012.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
