% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_logpost.R
\name{glm_logpost}
\alias{glm_logpost}
\title{The conditional and joint log posterior function}
\usage{
glm_logpost(Y, x, Params, callParam, splineArgs, priorArgs, Params_Transform)
}
\description{
The conditional and joint log posterior function
}
