% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_forecast.R
\name{individual_forecast}
\alias{individual_forecast}
\title{Function to predict forecast error measures from the
the fitted fformpp}
\usage{
individual_forecast(predicted, accmat = NULL, real.error = NULL,
  tslist = TRUE)
}
\arguments{
\item{predicted}{predicted forecast errors from predict_fformpp}

\item{accmat}{function to compute forecast accuracy}

\item{real.error}{optional, matrix of MASE values for all algorithms for test data}

\item{tslist}{list of time series, as in the format of Mcomp object}
}
\description{
This function is written based on the example codes in movingknots
package by Dr. Feng Li. Please see https://github.com/feng-li/movingknots for more details.
}
