% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_subsetsIdx.R
\name{knots_subsetsIdx}
\alias{knots_subsetsIdx}
\title{This function is to find the locations of subsets in the knots matrix}
\usage{
knots_subsetsIdx(knots.subsets, k, m)
}
\description{
This function is to find the locations of subsets in the knots matrix
}
