% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStepNewtonMove.R
\name{KStepNewtonMove}
\alias{KStepNewtonMove}
\title{This is the simple Newton move for spline models without dimession changes.}
\usage{
KStepNewtonMove(
  param.cur,
  gradhess.fun.name,
  KStep,
  Params,
  hessMethod,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\arguments{
\item{param.cur}{"matrix".
The initial values for the Newton update.}

\item{priorArgs}{priorArgs$prior_type:
priorArgs$ka0:
priorArgs$M:
priorArgs$S0:
priorArgs$n0:
priorArgs$mu0:
priorArgs$Sigma0:
priorArgs$ka.mu0:
priorArgs$ka.Sigma0:}
}
\value{
"list". See bellow.
\item   {gradObs.cur}
        {"matrix". The gradient}
\item   {invHessObs.cur}
        {"matrix". The inverse hessian matrix.}
\item   {param.cur}
        {"matrix". The updated paramters after K step Newton integrations.}
}
\description{
This is the simple Newton move for spline models without dimession changes.
}
\note{
First version: Wed Sep 29 17:18:22 CEST 2010;
      Current:       Wed Sep 29 17:18:34 CEST 2010.
TODO: Error handling when hessian is ill behaved.
      Think about how to add shrinkage in.
      And variable selection
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
