% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_IWishart.R
\name{linear_IWishart}
\alias{linear_IWishart}
\title{Calculate the posterior df and location matrix V from the inverse Wishart distribution.}
\usage{
linear_IWishart(Params, Y, x0, splineArgs, priorArgs, Params_Transform)
}
\description{
<details>
}
\note{
First version: Tue Feb 22 19:49:35 CET 2011;
      Current:       Tue Feb 22 19:49:42 CET 2011.
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
