% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat.x.delta.vecSigma4beta.inv.R
\name{Mat.x.delta.vecSigma4beta.inv}
\alias{Mat.x.delta.vecSigma4beta.inv}
\title{Compute a matrix to pre-multiply the gradient for vec beta's inverse variance}
\usage{
Mat.x.delta.vecSigma4beta.inv(Mat, delta4K.list, p, q, q.i)
}
\description{
Compute a matrix to pre-multiply the gradient for vec beta's inverse variance
}
