% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPropMain_rw.R
\name{MHPropMain_rw}
\alias{MHPropMain_rw}
\title{The Main file for Metropolis–Hastings algorithm}
\usage{
MHPropMain_rw(
  param.cur,
  gradhess.fun.name,
  logpost.fun.name,
  nNewtonStep,
  Params,
  hessMethod,
  Y.iCross,
  x.iCross,
  callParam,
  splineArgs,
  priorArgs,
  prop.df,
  Params_Transform,
  propMethod
)
}
\description{
The Main file for Metropolis–Hastings algorithm
}
