% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPropMain_sgld.R
\name{MHPropMain_sgld}
\alias{MHPropMain_sgld}
\title{The Main file for MCMC algorithm.}
\usage{
MHPropMain_sgld(
  param.cur,
  gradhess.fun.name,
  logpost.fun.name,
  nNewtonStep,
  Params,
  hessMethod,
  Y.iCross,
  x.iCross,
  callParam,
  splineArgs,
  priorArgs,
  algArgs,
  Params_Transform,
  propMethod
)
}
\description{
The Main file for MCMC algorithm.
}
