% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPropWithIWishart.R
\name{MHPropWithIWishart}
\alias{MHPropWithIWishart}
\title{Random walk Metropolis–Hastings algorithm for Sigma}
\usage{
MHPropWithIWishart(
  param.cur,
  logpost.fun.name,
  Params,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\description{
Random walk Metropolis–Hastings algorithm for Sigma
}
