% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPropWithKStepNewton.R
\name{MHPropWithKStepNewton}
\alias{MHPropWithKStepNewton}
\title{Metropolis–Hastings algorithm with K-step Newton method for the spline model.}
\usage{
MHPropWithKStepNewton(
  param.cur,
  gradhess.fun.name,
  logpost.fun.name,
  nNewtonStep,
  Params,
  hessMethod,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  prop.df,
  Params_Transform
)
}
\description{
Metropolis–Hastings algorithm with K-step Newton method for the spline model.
}
