% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params.subsets.R
\name{Params.subsets}
\alias{Params.subsets}
\title{Organize the subsets of the parameters by taking away the fixed parameters.}
\usage{
Params.subsets(p, splineArgs, Params_Fixed, Params_subsetsArgs)
}
\value{

}
\description{
<details>
}
\note{
First version: ; Current: .
}
\references{

}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
