% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fformpp.R
\name{predict_fformpp}
\alias{predict_fformpp}
\title{Function to predict forecast error measures from the
the fitted fformpp}
\usage{
predict_fformpp(model, feature.df, model.names, real.MASE = NULL,
  log = TRUE)
}
\arguments{
\item{model}{fitted fformpp models, output of the function fit_fformpp}

\item{feature.df}{feature matrix of test data}

\item{model.names}{vector of names of the forecast algorithms, similar to the order
of accmat argument in fit_ebmsr}

\item{real.MASE}{optional, matrix of MASE values for all algorithms for test data}

\item{log}{if log transformation is used to convert Y values to real line}
}
\description{
This function is written based on the example codes in movingknots
package by Dr. Feng Li. Please see https://github.com/feng-li/movingknots for more details.
}
