% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkMetropolis.R
\name{RandomWalkMetropolis}
\alias{RandomWalkMetropolis}
\title{Random walk Metropolis algorithm.}
\usage{
RandomWalkMetropolis(
  param.cur,
  logpost.fun.name,
  Params,
  Y,
  x0,
  callParam,
  splineArgs,
  priorArgs,
  Params_Transform
)
}
\description{
Random walk Metropolis algorithm.
}
