% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixmorm.R
\name{rmixnorm}
\alias{rmixnorm}
\title{Generate random variables from univariate and multivariate mixture normal distribution}
\usage{
rmixnorm(n, means, sigmas, weights)
}
\description{
Generate random variables from univariate and multivariate mixture normal distribution
}
